pyproject.toml
[tool.poetry]
name = "dspy-beyond-prompts"
version = "0.1.0"
description = "DSPy examples demonstrating modular AI programming"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
dspy = "^2.6.14"
openai = "^1.35.0"
anthropic = "^0.31.0"
pydantic = "^2.7.0"
python-dotenv = "^1.0.0"
requests = "^2.32.0"
aiohttp = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
black = "^24.4.0"
ruff = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
